requirements.txt
-i https://tech-pypi-prod.2u.com/simple/
python-docx
xlrd==0.9.2
xlwt==0.7.5
BeautifulSoup4
lxml
